django/valletpvbot_webhook/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
python-telegram-bot>=20.0
requests>=2.31.0
python-decouple>=3.8
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
celery>=5.3.4
redis>=5.0.1
